+++ test/union-test.js
@@ -7,30 +7,49 @@ describe('mapshaper-union.js', function () {
   it('union of two polygons with identical fields adds disambiguating suffixes', function(done) {
     var fileA = 'test/test_data/issues/union/polygonA.json';
     var fileB = 'test/test_data/issues/union/polygonB.json';
-    var cmd = `-i ${fileA} -union ${fileB} name=merged -o`;
+    var cmd = `-i ${fileA} ${fileB} combine-files -union name=merged -o`;
     api.applyCommands(cmd, {}, function(err, out) {
       var features = JSON.parse(out['merged.json']).features;
       var records = _.pluck(features, 'properties');
       assert.deepEqual(records, [
-        { name_B: 'B', value_B: 8, name_A: null, value_A: null },
-        { name_B: 'B', value_B: 8, name_A: 'A', value_A: 4 },
-        { name_B: null, value_B: null, name_A: 'A', value_A: 4 }
+        { name_2: null, value_2: null, name_1: 'A', value_1: 4 },
+        { name_2: 'B', value_2: 8, name_1: 'A', value_1: 4 },
+        { name_2: 'B', value_2: 8, name_1: null, value_1: null }
       ]);
       done();
     });
   });
 
-  it('-union add-fid adds FID_A and FID_B fields', function(done) {
+  it('fields= option selects fields to retain', function(done) {
     var fileA = 'test/test_data/issues/union/polygonA.json';
     var fileB = 'test/test_data/issues/union/polygonB.json';
-    var cmd = `-i ${fileA} -union add-fid ${fileB} name=merged -o`;
+    var cmd = `-i ${fileA} ${fileB} combine-files -union fields=name name=merged -o`;
     api.applyCommands(cmd, {}, function(err, out) {
       var features = JSON.parse(out['merged.json']).features;
       var records = _.pluck(features, 'properties');
       assert.deepEqual(records, [
-        { FID_A: -1, FID_B: 0, name_B: 'B', value_B: 8, name_A: null, value_A: null },
-        { FID_A: 0, FID_B: 0, name_B: 'B', value_B: 8, name_A: 'A', value_A: 4 },
-        { FID_A: 0, FID_B: -1, name_B: null, value_B: null, name_A: 'A', value_A: 4 }
+        { name_2: null, name_1: 'A'},
+        { name_2: 'B', name_1: 'A'},
+        { name_2: 'B', name_1: null}
+      ]);
+      done();
+    });
+  });
+
+  it('union of three polygons with identical fields', function(done) {
+    var fileA = 'test/test_data/issues/union/polygonA.json';
+    var fileB = 'test/test_data/issues/union/polygonB.json';
+    var fileC = 'test/test_data/issues/union/polygonC.json';
+    var cmd = `-i ${fileA} ${fileB} ${fileC} combine-files -union name=merged -o`;
+    api.applyCommands(cmd, {}, function(err, out) {
+      var features = JSON.parse(out['merged.json']).features;
+      var records = _.pluck(features, 'properties');
+      assert.deepEqual(records, [
+        { name_2: null, value_2: null, name_1: 'A', value_1: 4, name_3: null, value_3: null },
+        { name_2: 'B', value_2: 8, name_1: 'A', value_1: 4, name_3: null, value_3: null  },
+        { name_2: 'B', value_2: 8, name_1: null, value_1: null, name_3: null, value_3: null  },
+        { name_2: 'B', value_2: 8, name_1: null, value_1: null, name_3: 'C', value_3: 0  },
+        { name_2: null, value_2: null, name_1: null, value_1: null, name_3: 'C', value_3: 0  }
       ]);
       done();
     });
@@ -39,14 +58,14 @@ describe('mapshaper-union.js', function () {
   it('union of two polygons with different fields preserves field names', function(done) {
     var fileA = 'test/test_data/issues/union/polygonA.json';
     var fileB = 'test/test_data/issues/union/polygonB.json';
-    var cmd = `-i ${fileA} -rename-fields nameA=name,valueA=value -union ${fileB} name=merged -o`;
+    var cmd = `-i ${fileA} -rename-fields nameA=name,valueA=value -i ${fileB} -union target=* name=merged -o`;
     api.applyCommands(cmd, {}, function(err, out) {
       var features = JSON.parse(out['merged.json']).features;
       var records = _.pluck(features, 'properties');
       assert.deepEqual(records, [
-        { name: 'B', value: 8, nameA: null, valueA: null },
+        { name: null, value: null, nameA: 'A', valueA: 4 },
         { name: 'B', value: 8, nameA: 'A', valueA: 4 },
-        { name: null, value: null, nameA: 'A', valueA: 4 }
+        { name: 'B', value: 8, nameA: null, valueA: null }
       ]);
       done();
     });
@@ -58,14 +77,14 @@ describe('mapshaper-union.js', function () {
       type: 'Polygon',
       coordinates: [[[1, 1], [2, 2], [3, 1], [2, 0], [1, 1]]]
     }
-    var cmd = `-i polygonB.json -i ${fileA} -union polygonB name=merged -o`;
+    var cmd = `-i polygonB.json -i ${fileA} -union target=polygonA,polygonB name=merged -o`;
     api.applyCommands(cmd, {'polygonB.json': geomB}, function(err, out) {
       var features = JSON.parse(out['merged.json']).features;
       var records = _.pluck(features, 'properties');
       assert.deepEqual(records, [
-        { name: null, value: null },
         { name: 'A', value: 4 },
-        { name: 'A', value: 4 }
+        { name: 'A', value: 4 },
+        { name: null, value: null }
       ]);
       done();
     });